:recycle: split activity pages
