+++ server/vitest.pure.config.js
@@ -7,12 +7,14 @@ export default defineConfig({
     include: [
       "tests/unit/chartJsCartesianCompiler.test.js",
       "tests/unit/embeddedChartPayload.test.js",
+      "tests/unit/kpiReview.test.js",
       "tests/unit/nodemail.passwordReset.test.js",
       "tests/unit/observationPeriodCorpus.test.js",
       "tests/unit/observations.test.js",
       "tests/unit/periodEvaluator.test.js",
       "tests/unit/periodObservationPersistence.test.js",
       "tests/unit/replaySynthbrewPeriods.test.js",
+      "tests/unit/scheduleWeekdays.test.js",
       "tests/unit/visualizationCompilers.test.js",
     ],
     maxWorkers: 1,
